Refactor monolithic script into modules
